requirements.txt
feedparser>=6.0.0
requests>=2.31.0
PyYAML>=6.0
python-dateutil>=2.8.2
